flujo completo

+++ gamat-backend/src/main/java/usach/cl/gamatbackend/entities/Item.java
@@ -170,7 +170,7 @@ public class Item  implements Serializable{
 		return totalPrice;
 	}
 
-	public void setTotalPrice(int totalPrice) {
+	public void setTotalPrice(long totalPrice) {
 		this.totalPrice = totalPrice;
 	}
